resolve dependencies without using Coursier
Since we do not need to reconcile version information, not using
coursier makes lots of things simpler. It allows for more stringent
types, so that the resolutions are more evidently correct. In
particular, we do not need to bother with all the kinds of metadata of
Coursier Dependency types that introduce nasty ambiguities.