fix missing handling of download errors when downloading metadata
When the connection to the metadata channel could not be established, this
could lead to a crash (especially, if the channel contents file is
cached, but the metadata JSON files cannot be accessed).

For the GUI, this meant that the update process got stuck.

Now, download errors are properly handled for metadata files (similarly
to how they were handled for asset files already).